fix: disable i18n plugin as default on gunshi
