+++ packages/gunshi/src/constants.ts
@@ -33,6 +33,5 @@ export const COMMAND_OPTIONS_DEFAULT: CliOptions<DefaultGunshiParams> = {
   renderHeader: undefined,
   renderUsage: undefined,
   renderValidationErrors: undefined,
-  translationAdapterFactory: undefined,
   plugins: undefined
 }
