+++ packages/resources/package.json
@@ -72,7 +72,6 @@
   "devDependencies": {
     "@types/node": "catalog:",
     "deno": "catalog:",
-    "gunshi": "workspace:*",
     "jsr": "catalog:",
     "jsr-exports-lint": "catalog:",
     "publint": "catalog:",
